Feature/dashboard feature (#47)
* feat: Add dashboard page with role-based views

This commit introduces a new dashboard page that dynamically renders content
based on the user's role (merchant, admin, or clerk). It also includes
the installation and configuration of Tailwind CSS for styling.

The primary goal is to provide a central, role-specific hub for each user type.
The merchant view displays overall sales and store performance data,
the admin view manages pending requests, and the clerk view handles inventory
management.

* feat: add dashboard features for merchant, admin and clerk

* Create component with card

* [200~feat: Refactor dashboard into three distinct user roles

* complete dashboard features

* feat : authentication for role based dashboard complete

* feat : authentication for role based dashboard complete

* Add crud features to admin dashboard

* Add category feature  to clerk dashboard

* Add category feature  to merchant dashboard

* feat: Connect client jsx code to server endpoints

* feat: Connect client jsx code to server endpoints

* feat: Connect client jsx code to server endpoints